Broaden compability for static linking
In recent 0.3.x changes, support for static linking was added (thanks!).
It doesn't seem to work with the version of `ld` used with clang on mac.

This commit changes the static linking format to be more general, and to
remove the PCRE path hardcoding. I'm not an expert on this, and I can't
say whether this is definitively better across different build tools.
However, it works on mac!